+++ src/main/java/org/libreoffice/ci/gerrit/buildbot/commands/VerifyCommand.java
@@ -9,36 +9,47 @@
 
 package org.libreoffice.ci.gerrit.buildbot.commands;
 
+import java.lang.reflect.Field;
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Set;
 
 import org.kohsuke.args4j.Argument;
 import org.kohsuke.args4j.Option;
 import org.libreoffice.ci.gerrit.buildbot.config.BuildbotConfig;
-import org.libreoffice.ci.gerrit.buildbot.config.BuildbotProject;
 import org.libreoffice.ci.gerrit.buildbot.logic.BuildbotLogicControl;
 import org.libreoffice.ci.gerrit.buildbot.model.GerritJob;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.google.gerrit.common.data.ApprovalType;
 import com.google.gerrit.common.data.ApprovalTypes;
 import com.google.gerrit.common.data.GlobalCapability;
 import com.google.gerrit.extensions.annotations.RequiresCapability;
+import com.google.gerrit.reviewdb.client.ApprovalCategory;
+import com.google.gerrit.reviewdb.client.ApprovalCategoryValue;
 import com.google.gerrit.reviewdb.client.Change;
 import com.google.gerrit.reviewdb.client.PatchSet;
 import com.google.gerrit.reviewdb.client.RevId;
 import com.google.gerrit.reviewdb.server.ReviewDb;
+import com.google.gerrit.server.IdentifiedUser;
+import com.google.gerrit.server.patch.PublishComments;
 import com.google.gerrit.server.project.ProjectControl;
 import com.google.gerrit.sshd.SshCommand;
+import com.google.gerrit.util.cli.CmdLineParser;
 import com.google.gwtorm.server.OrmException;
 import com.google.gwtorm.server.ResultSet;
 import com.google.inject.Inject;
 
 @RequiresCapability(GlobalCapability.VIEW_QUEUE)
-public final class ScheduleCommand extends SshCommand {
-	static final Logger log = LoggerFactory.getLogger(ScheduleCommand.class);
+public final class VerifyCommand extends SshCommand {
+	static final Logger log = LoggerFactory.getLogger(VerifyCommand.class);
 
+	@Inject
+	IdentifiedUser user;
+	
 	@Inject
 	BuildbotLogicControl control;
 
@@ -48,15 +59,29 @@ public final class ScheduleCommand extends SshCommand {
 	@Inject
 	private ApprovalTypes approvalTypes;
 
+    @Inject
+    private PublishComments.Factory publishCommentsFactory;
+
 	@Inject
 	private ReviewDb db;
-
+	
 	private final Set<PatchSet.Id> patchSetIds = new HashSet<PatchSet.Id>();
 
-	@Option(name = "--project", aliases = { "-p" }, required = true, metaVar = "PROJECT", usage = "name of the project for which the job should be scheduled")
+	private List<ApproveOption> optionList;
+	
+	@Override
+	protected final CmdLineParser newCmdLineParser(Object options) {
+		final CmdLineParser parser = super.newCmdLineParser(options);
+		for (ApproveOption c : optionList) {
+			parser.addOption(c, c);
+		}
+		return parser;
+	}
+	
+	@Option(name = "--project", aliases = { "-p" }, required = true, metaVar = "PROJECT", usage = "name of the project for patch set to be verified")
 	private ProjectControl projectControl;
 
-	@Argument(index = 0, required = true, multiValued = true, metaVar = "{COMMIT | CHANGE,PATCHSET}", usage = "commit(s) or patch set(s) to schedule")
+	@Argument(index = 0, required = true, multiValued = false, metaVar = "{COMMIT | CHANGE,PATCHSET}", usage = "commit or patch set to verify")
 	void addPatchSetId(final String token) {
 		try {
 			patchSetIds.addAll(parsePatchSetId(token));
@@ -73,7 +98,7 @@ public final class ScheduleCommand extends SshCommand {
 
 	@Override
 	public void run() throws UnloggedFailure, Failure, Exception {
-		log.debug("schedule");
+		log.debug("verify");
 		
 		final String p = projectControl.getProject().getName();
 		if (!config.isProjectSupported(p)) {
@@ -83,67 +108,74 @@ public final class ScheduleCommand extends SshCommand {
 			stderr.print(tmp + "\n");
 			return;
 		}
+		if (patchSetIds.size() > 1) {
+			String tmp = "error: only one commit|patch set can be provided";
+			log.warn(tmp);
+			stderr.print(tmp + "\n");
+			return;
+		}
 		for (PatchSet.Id id : patchSetIds) {
-			doSchedule(id);
+			doVerify(id);
 		}
 	}
 
-	private void doSchedule(PatchSet.Id id) throws OrmException {
+	private void doVerify(PatchSet.Id id) throws OrmException {
 		PatchSet patchSet = db.patchSets().get(id);
-
-		Change change = db.changes().get(id.getParentKey());
-
-		// this value seems not to be set
-		// only check if value is set (new gerrit release)
-		if (change.getStatus() != null) {
-			// check if the patch set is NEW
-			if (Change.Status.NEW != change.getStatus()) {
-				String tmp = String.format(
-						"error: status of patch set %s is not NEW",
-						id.toString());
-				log.warn(tmp);
-				stderr.print(tmp + "\n");
-				return;
-			}
-		}
-
 		// check if build job is pending for this patch set
 		GerritJob job = control.findJobByRevision(projectControl.getProject().getName(),
 				patchSet.getRevision()
 				.toString());
 		if (job != null) {
 			String tmp = String.format(
-					"error: build job already scheduled at %s",
+					"error: build job still pending from %s",
 					job.getStartTime());
 			log.warn(tmp);
 			stderr.print(tmp + "\n");
 			return;
 		}
+		overrideVerifyStatus(patchSet);
+	}
 
-		BuildbotProject p = config.findProject(projectControl.getProject().getName());
-		// check branch if configured
-		if (!p.getBranches().isEmpty()) {
-			if (!p.getBranches().contains(change.getDest().getShortName())) {
-				String tmp = String.format("error: branch not match  %s",
-						change.getDest().getShortName());
-				log.warn(tmp);
-				stderr.print(tmp + "\n");
-				return;
-			}
-		}
+    public void overrideVerifyStatus(PatchSet patchSet) {
+    	Set<ApprovalCategoryValue.Id> aps = new HashSet<ApprovalCategoryValue.Id>();
+    	Short v = 0;
+        for (ApproveOption ao : optionList) {
+          v = ao.value();
+          if (v != null) {
+            aps.add(new ApprovalCategoryValue.Id(ao.getCategoryId(), v));
+          }
+        }
+        
+        try {
+			StringBuilder builder = new StringBuilder(
+					String.format(
+							"Verification Status of the Buildbot is manually set to %d by %s",
+							v, user.getUserName()));
+			getCommenter(aps, patchSet, builder).call();
+        } catch (Exception e) {
+            e.printStackTrace();
+            die(e);
+        }
+    }
 
-		synchronized (control) {
-			 log.debug("dispatch event branch: {}, ref: {}",
-                     change.getDest().getShortName(),
-                     patchSet.getRefName());
-             control.startGerritJob(projectControl.getProject().getName(), change, patchSet);
+	private PublishComments getCommenter(
+			Set<ApprovalCategoryValue.Id> aps, PatchSet patchset,
+			StringBuilder builder) throws NoSuchFieldException,
+			IllegalAccessException {
+		PublishComments commenter = publishCommentsFactory.create(patchset.getId(),
+		        builder.toString(), aps, true);
+		if (config.isForgeReviewerIdentity()) {
+			// Replace current user with buildbot user
+			Field field = commenter.getClass().getDeclaredField("user");
+			field.setAccessible(true);
+			field.set(commenter, control.getBuildbot());
 		}
+		return commenter;
 	}
 
 	private Set<PatchSet.Id> parsePatchSetId(final String patchIdentity)
 			throws UnloggedFailure, OrmException {
 		// By commit?
-		//
 		if (patchIdentity.matches("^([0-9a-fA-F]{4," + RevId.LEN + "})$")) {
 			final RevId id = new RevId(patchIdentity);
 			final ResultSet<PatchSet> patches;
@@ -212,4 +244,29 @@ public final class ScheduleCommand extends SshCommand {
 	private static UnloggedFailure error(final String msg) {
 		return new UnloggedFailure(1, msg);
 	}
+	
+	@Override
+	protected void parseCommandLine() throws UnloggedFailure {
+		optionList = new ArrayList<ApproveOption>();
+
+		for (ApprovalType type : approvalTypes.getApprovalTypes()) {
+			String usage = "";
+			final ApprovalCategory category = type.getCategory();
+			if (!category.getName().equals("Verified")) {
+				continue;
+			}
+			usage = "score for " + category.getName() + "\n";
+
+			for (ApprovalCategoryValue v : type.getValues()) {
+				usage += v.format() + "\n";
+			}
+
+			final String name = "--"
+					+ category.getName().toLowerCase().replace(' ', '-');
+			optionList.add(new ApproveOption(name, usage, type));
+		}
+
+		super.parseCommandLine();
+	}
+
 }